Adjusted the Controllers and Services of the Entities
Adjusted the Controllers and Services of the Entities for handling relationships

+++ backend/SPRINGBOOT_PROJECT/swiftthrift/swiftthrift/src/main/java/edu/cit/swiftthrift/service/StoreRatingService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class StoreRatingService {
@@ -13,32 +14,36 @@ public class StoreRatingService {
     @Autowired
     private StoreRatingRepository storeRatingRepository;
 
-    
-    public List<StoreRating> getAllProductRatings() {
+    public List<StoreRating> getAllStoreRatings() {
         return storeRatingRepository.findAll();
     }
 
-    public StoreRating getProductRatingById(int id) {
+    public StoreRating getStoreRatingById(int id) {
         return storeRatingRepository.findById(id).orElse(null);
     }
 
-    public StoreRating createProductRating(StoreRating productRating) {
-        return storeRatingRepository.save(productRating);
+    public StoreRating createStoreRating(StoreRating storeRating) {
+        return storeRatingRepository.save(storeRating);
     }
 
-    public StoreRating updateProductRating(int id, StoreRating productRating) {
-        StoreRating existingRating = storeRatingRepository.findById(id).orElse(null);
+    public StoreRating updateStoreRating(int id, StoreRating storeRating) {
+        Optional<StoreRating> existingRatingOpt = storeRatingRepository.findById(id);
     
-        if (existingRating != null) {
-            existingRating.setName(productRating.getName());
-            existingRating.setRating(productRating.getRating());
-            existingRating.setDate(productRating.getDate());
+        if (existingRatingOpt.isPresent()) {
+            StoreRating existingRating = existingRatingOpt.get();
+            existingRating.setName(storeRating.getName());
+            existingRating.setRating(storeRating.getRating());
+            existingRating.setDate(storeRating.getDate());
             return storeRatingRepository.save(existingRating);
         }
         return null; 
     }
 
-    public void deleteProductRating(int id) {
-        storeRatingRepository.deleteById(id);
+    public boolean deleteStoreRating(int id) {
+        if (storeRatingRepository.existsById(id)) {
+            storeRatingRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
